refactor: 예약 step2 activity fragment로 변경
