+++ app/src/main/AndroidManifest.xml
@@ -61,9 +61,6 @@
             android:name=".ui.reservation.ReservationStep3Activity"
             android:exported="false" />
         <activity
-            android:name=".ui.reservation.ReservationStep2Activity"
-            android:exported="false" />
-        <activity
             android:name=".ui.main.companion.LiveCompanionActivity"
             android:exported="false" />
         <activity
