+++ app/src/main/res/layout/fragment_reservation_step2.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.reservation.ReservationStep2Activity">
+    tools:context=".ui.reservation.ReservationStep2Fragment">
 
     <ImageView
         android:id="@+id/back_btn"
@@ -21,7 +21,7 @@
         android:id="@+id/select_service_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
+        android:layout_marginStart="30dp"
         android:layout_marginTop="8dp"
         android:text="@string/select_service"
         android:textSize="18sp"
